Add tests for ask helper
We're not sure if this will be useful, but it might be nice to expose a
helper that allows Dock configuration files to ask the user for
confirmation before performing some actions.

The `ask` helper is TTY aware, so it won't display the
questions if we aren't in an interactive context.