Add help flag for documentation
We include documentation in the executable itself and display it by
searching for magic strings and filtering.

Unfortunately this requires we use `grep` and `sed`. This seems fine
since most users will be running this on macOS or "normal" Linux machine
(i.e. not an Docker image) but if we hit problems we can look into
implementing a pure Bash solution.