Rename hostname option to container_hostname
Setting the --hostname flag of the `docker run` command is incompatible
with a number of network-related options (e.g. --network=host). If we
want to be able to network two Dock containers, we can't set this flag
by default.

While here, we renamed the option `container_hostname` to avoid issues
with the `hostname` executable, as this will surely cause confusion.