Remove docker-compose verification during extension
docker-compose file verification shouldn't occur during the extension step since
the docker-compose files are meant to resolve in the context of the dock
container (e.g. env vars are set within and only within the dock container) so it
doesn't really make sense to do it before the dock container is even created.