Refactor exposed port identification
The previous mechanism for identifying exposed ports on
host machines was flakey and not really an ideal/efficient
means of identifying exposed ports (i.e attempting to connect
to port with a 1 second timeout). This change modifies the method
to make use of netstat and varies the usage based on the host
OS appropriately.