Add startup_services dock configuration functionality
This change enables owners of projects to specify which
services, defined within the project's docker-compose file,
to launch when standing up the project within an extended dock
environment. Particulary of note here is that this change allows
project owners to target variations of services and backend
components depending on context.