Set project dock repo root to current working directory
It should be safe to assume that dock can operate whereever it
is executed as long as the current directory meets the general
configuration requirements (i.e. is a part of a git repository
and contains a .dock file - for at least specifying how the Dock
development container is constructed).

So, rather than default a project's repo root to the top level
git repository root, default to the current working directory and trust
that it and the user adhere to the previously mentioned requirements.